backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
httpx
supabase
openai
